DocFix
Some proposed tidying of README file. Because it’s Markdown I’ve turned
all the links into clickable links, which tidies them up a bit and
saves space. I’ve also tweaked the capitalisation to match each other,
and I’ve tweaked the Homebrew instructions to remove the unnecessary
invocation of ``` --Framework ``` (Homebrew’s Python builds a Framework
naturally, it doesn’t need invoking).

Not trying to be an arse here; If you like the document how it is, I
apologise for proposing changes to it, but hopefully these proposed
changes will be useful to you.